+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Change Log
 
 ## WIP
-### Improved
+### Added
 - [KV] range functions automatically spill to disk when results do not fit in
-  memory. The on-disk temporary result file behaves like an `IPersistentVector`,
-  so there is no API difference from the in-memory results. #108
+  memory. The results as a whole still behave like an `IPersistentVector`,
+  so there is no API level change from the user's point of view. #108
+- [KV] `:spill-threshold` option, in MB, to control the threshold of available free
+  memory, below which spill-to-disk will be triggered.
+- [KV] `:spill-interval` option, the number of result items added before a check
+  for spill should be performed.
 
 ## 0.6.29
 ### Added